Now edit event has better UI
