+++ static/editEvent.css
@@ -1,0 +1,75 @@
+body{
+  display: flex;
+  overflow: auto;
+  flex-direction: column;
+  margin: 0;
+  min-height: 100vh;
+  background-image: url(images/background.jpg);
+  background-attachment: fixed;
+  background-position: center;
+  background-repeat: no-repeat;
+  background-size: cover;
+}
+
+
+h1 {color: #222; margin-top: -20px; margin-bottom: 20px;}
+label {
+  color: #222;
+  font-weight: 700;
+  font-size: 13px;
+}
+
+.content{
+  background-color:rgba(255,255,255,.8);
+  border-radius:.25em;
+  box-shadow:0 0 .25em rgba(0, 0, 0, 0.25);
+  box-sizing:border-box;
+  left:50%;
+  padding: 5vmin 10vmin;
+  position: absolute;
+  text-align: center;
+  top:50%;
+  transform:translate(-50%, -50%);
+  height: 80%;
+  width: 50%;
+  overflow: auto;
+}
+
+#submit_btn{
+    background-color: #222;
+    border: 1px solid #222;
+    color: #9d9d9d;
+    border-radius: 5px;
+    width: 40%;
+    height: 15%;
+    font-size: 30px;
+}
+
+#submit_btn:hover{
+    background-color: #222;
+    color: white;
+}
+
+.group{
+    margin-bottom: 15px;
+}
+
+input{
+    font-size: 15px;
+}
+
+textarea {
+  width: 100%;
+  height: 150px;
+  padding: 12px 20px;
+  box-sizing: border-box;
+  border: 2px solid #ccc;
+  border-radius: 4px;
+  background-color: #f8f8f8;
+  font-size: 15px;
+  resize: none;
+}
+
+#location {
+  width: 100%;
+}
