+++ templates/editEvent.html
@@ -2,6 +2,11 @@
 <html>
     <head>
         <title>Edit your Event</title>
+        <meta name="viewport" content="width=device-width, initial-scale=1">
+        <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.4.1/css/bootstrap.min.css">
+        <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.5.1/jquery.min.js"></script>
+        <script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.4.1/js/bootstrap.min.js"></script>
+        <link rel="stylesheet" type="text/css" href="{{ url_for('static', filename='editEvent.css') }}">
         <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.5.1/jquery.min.js"></script>
 
         <script type="text/javascript">
@@ -41,7 +46,18 @@
         </script>
     </head>
     <body>
-      <h1> Sports Eventer </h1>
+      <nav class="navbar navbar-inverse" style="font-size: 20px;">
+      <div class="container-fluid">
+        <div class="navbar-header">
+          <a class="navbar-brand" href="/" style="font-size: 25px;">SportsEventer</a>
+        </div>
+        <ul class="nav navbar-nav navbar-right">
+            <li><a href="/newEvent"><span class="glyphicon glyphicon-plus"></span> Create an Event</a></li>
+            <li><a href="/myProfile"><span class="glyphicon glyphicon-user"></span> My Profile</a></li>
+            <li><a href="/logout"><span class="glyphicon glyphicon-log-out"></span> Log out</a></li>
+        </ul>
+      </div>
+    </nav>
       <header>
           <button type="button" onclick="window.location='/'">Home</button>
       </header>
