+++ templates/myProfile.html
@@ -166,8 +166,6 @@
         </tr>
 
       </table>
-      <br><br>
-      <a id="edit_profile" href="/myProfile/{{user.user_id}}/edit">Edit Profile</a>
       <h3>Events I participate in</h3>
       <table id="joined_events">
         <tr>
@@ -183,6 +181,8 @@
           <th>Descriptions</th>
         </tr>
       </table>
+      <br><br>
+      <a id="edit_profile" href="/myProfile/{{user.user_id}}/edit">Edit Profile</a>
     </div>
     </div>
