+++ user.py
@@ -1,7 +1,6 @@
 from flask import Flask
 from werkzeug.security import generate_password_hash, check_password_hash
 from db_config import mydb
-from event import Event
 
 
 cursor = mydb.cursor(buffered=True)
@@ -44,22 +43,6 @@ class User:
         else:
             return User(*row)
 
-    @staticmethod
-    def get_joined_events(user_id):
-        query = 'SELECT event_id FROM UsersEvents WHERE user_id=%s'
-        value = (user_id,)
-        cursor.execute(query, value)
-        results = cursor.fetchall()
-        events = []
-        if results is None:
-            return;
-        else:
-            for row in results:
-                events.append(Event.find(*row))
-            if len(events) == 0:
-                return;
-            else:
-                return events
 
 
     @staticmethod
